💥 [Fonts]: Font rendering changes
Fonts were rendering at larger than normal sizes. This change
ensures that all glpyhs are rendered at their correct size. This
was needed to allow sensible generations of bounding boxes for
ASGE:Text instances. To support this a new function has been added
that calculates the distance in the Y axis min-max from the
originally requested baseline. See `boundsY` on GLFontSet.

Note: This commit will potentially effet the size of fonts being
rendered in existing games. However, as this is not a API breaking
change it has been released as a X.Y bumped version.

+++ engine/src/Engine/OpenGL/GLAtlasManager.cpp
@@ -134,8 +134,7 @@ int ASGE::GLAtlasManager::createAtlas(FT_Face& face, const char* name, int pt)
 
   // store the line spacing
   set.line_height =
-    static_cast<int>(
-      (face->size->metrics.ascender - face->size->metrics.descender)) / 64;
+      static_cast<float>(face->size->metrics.ascender - face->size->metrics.descender) / 64.0F;
 
   FT_Done_Face(face);
